Split test module out
